Step 25

+++ demo/src/main/java/com/example/demo/controller/TodolistController.java
@@ -79,6 +79,9 @@ public class TodolistController {
 	
 	@RequestMapping(value="/delete-todo",method = RequestMethod.GET)
 	public String deleteTodo(@RequestParam int id) {
+		if(id==1)
+			throw new RuntimeException("Something went wrong");
+		
 		todoService.deleteTodo(id);
 		
 		return "redirect:/todolist";
